+++ R/net.R
@@ -3,10 +3,10 @@
 #' This dataset is a SpatialLineDataFrame of sp package with roads
 #' from a traffic simulations made by CET Sao Paulo, Brazil
 #'
-#' @format A data frame with 1796 rows and 1 variables:
+#' @format A Spatial data.frame (sf) with 1796 rows and 1 variables:
 #' \describe{
-#'   \item{ldv}{Light Duty Vehicles (1/h)}
-#'   \item{hdv}{Heavy Duty Vehicles (1/h)}
+#'   \item{ldv}{Light Duty Vehicles (veh/h)}
+#'   \item{hdv}{Heavy Duty Vehicles (veh/h)}
 #'   \item{lkm}{Length of the link (km)}
 #'   \item{ps}{Peak Speed (km/h)}
 #'   \item{ffs}{Free Flow Speed (km/h)}
@@ -14,9 +14,9 @@
 #'   \item{lanes}{Number of lanes per link}
 #'   \item{capacity}{Capacity of vehicles in each link (1/h)}
 #'   \item{tmin}{Time for travelling each link (min)}
+#'   \item{geometry}{geometry}
 #' }
 #' @source \url{http://www.cetsp.com.br/}
 #' @usage data(net)
 #' @docType data
-#' @import sp
 "net"